refactor(control): split udp flow
