+++ crates/honk-core/src/control/connection/mod.rs
@@ -24,6 +24,8 @@ pub(super) use context::{ConnectionGuard, ControlPlaneHandle};
 
 mod tcp;
 
+mod udp;
+
 mod flow;
 
 #[cfg(test)]
